split_sbp
